Medpages

+++ Pharmacy/Pharmacy.Core/PharmacyDbContextExtension.cs
@@ -168,7 +168,7 @@ namespace Pharmacy.Core
                 Name = "Синупрет табл. в/о №50",
                 Code = "4882",
                 Dosage = "",
-                Price = 125.62,
+                Price = (float)125.62,
                 ReleaseForm = "таблетки для внутрішнього застосування",
                 PhotoPath = "https://i.apteka24.ua/products/8986bcef-7cf8-4894-854a-825e8f724920.jpeg",                
             };
